Refactored attributes with Double reference type to double

+++ src/main/java/com/kuna_backend/models/Product.java
@@ -31,7 +31,7 @@ public class Product {
     @Column (name = "desc_product", length = 2048, nullable = false)
     private String description;
     @Column (name = "price", nullable = false)
-    private Double price;
+    private double price;
     @Column (name = "image_url", length = 256, nullable = false)
     private String imageUrl;
     @Column (name = "created_at")
@@ -57,7 +57,7 @@ public class Product {
         productVariations = new ArrayList<>();
     }
 
-    public Product(Integer id, String name, String description, Double price, String imageUrl, LocalDateTime createdAt, LocalDateTime modifiedAt, Category category) {
+    public Product(long id, String name, String description, double price, String imageUrl, LocalDateTime createdAt, LocalDateTime modifiedAt, Category category) {
         this.id = id;
         this.name = name;
         this.description = description;
@@ -73,7 +73,7 @@ public class Product {
         productVariations = new ArrayList<>();
     }
 
-    public Product(Integer productId, ProductDto productDto, Category category) {
+    public Product(long productId, ProductDto productDto, Category category) {
         productVariations = new ArrayList<>();
     }
 
@@ -101,11 +101,11 @@ public class Product {
         this.description = description;
     }
 
-    public Double getPrice() {
+    public double getPrice() {
         return price;
     }
 
-    public void setPrice(Double price) {
+    public void setPrice(double price) {
         this.price = price;
     }
